third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_constructor-expected.txt

This is a testharness.js-based test.
[FAIL] getCoalescedEvents in event
  assert_equals: expected false but got true
Harness: the test ran to completion.
